Move AuthorizationType to core
