+++ aws-api/src/main/java/com/amplifyframework/api/aws/AWSApiPluginConfigurationReader.java
@@ -17,6 +17,7 @@ package com.amplifyframework.api.aws;
 
 import com.amplifyframework.AmplifyException;
 import com.amplifyframework.api.ApiException;
+import com.amplifyframework.api.AuthorizationType;
 
 import org.json.JSONException;
 import org.json.JSONObject;
