+++ aws-api/src/main/java/com/amplifyframework/api/aws/ApiConfiguration.java
@@ -19,6 +19,8 @@ import android.annotation.SuppressLint;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import com.amplifyframework.api.AuthorizationType;
+
 import java.util.Objects;
 
 /**
