+++ aws-api/src/main/java/com/amplifyframework/api/aws/AppSyncSigV4SignerInterceptorFactory.java
@@ -17,6 +17,7 @@ package com.amplifyframework.api.aws;
 
 import com.amplifyframework.AmplifyException;
 import com.amplifyframework.api.ApiException;
+import com.amplifyframework.api.AuthorizationType;
 import com.amplifyframework.api.aws.sigv4.ApiKeyAuthProvider;
 import com.amplifyframework.api.aws.sigv4.AppSyncSigV4SignerInterceptor;
 import com.amplifyframework.api.aws.sigv4.CognitoUserPoolsAuthProvider;
