+++ aws-api/src/main/java/com/amplifyframework/api/aws/sigv4/AppSyncSigV4SignerInterceptor.java
@@ -17,7 +17,7 @@ package com.amplifyframework.api.aws.sigv4;
 
 import androidx.annotation.NonNull;
 
-import com.amplifyframework.api.aws.AuthorizationType;
+import com.amplifyframework.api.AuthorizationType;
 import com.amplifyframework.api.aws.EndpointType;
 import com.amplifyframework.util.Empty;
